requirements.txt
torch
torchvision
matplotlib
scikit-learn
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0